lr_3/requirements.txt
fastapi
pydantic
uvicorn[standard]
sqlalchemy
pandas
psycopg2-binary
redis
pymongo
pyjwt
python-jose[cryptography] 
passlib[bcrypt]
python-multipart

